refactor: move DefaultCmsModelOpenSearchIndex and provider to utils-os
Both ddb-es and pg-os had identical copies of DefaultCmsModelOpenSearchIndex
and DefaultCmsModelOpenSearchIndexProvider. Moved to utils-os as the single
source of truth. Both packages now re-export from utils-os.
