--- packages/api-headless-cms-ddb-es/src/configurations.ts
@@ -1,5 +1,0 @@
-export {
-    createConfigurations,
-    type Configurations,
-    type CmsElasticsearchParams
-} from "@webiny/api-headless-cms-utils-os/configurations.js";
