+++ packages/api-headless-cms-ddb-es/src/features/CmsModelOpenSearchIndex/abstractions.ts
@@ -2,6 +2,6 @@ export {
     CmsModelOpenSearchIndex,
     type ICmsModelOpenSearchIndex,
     type ICmsModelOpenSearchIndexParams,
-    type ICmsModelOpenSearchIndexResult,
-    type StorageCmsModel
+    type ICmsModelOpenSearchIndexResult
 } from "@webiny/api-headless-cms-utils-os/features/CmsModelOpenSearchIndex/abstractions.js";
+export { type StorageCmsModel } from "@webiny/api-headless-cms/types/index.js";
