+++ packages/api-headless-cms-ddb-es/src/operations/entry/DdbEsCreateEntryRevisionFrom.ts
@@ -16,7 +16,7 @@ import {
 } from "@webiny/api-headless-cms-utils-os/exports/api/cms/opensearch.js";
 import { CompressionHandler } from "@webiny/utils/exports/api.js";
 import { CmsModelOpenSearchIndexProvider } from "~/features/CmsModelOpenSearchIndex/index.js";
-import { createConfigurations } from "~/configurations.js";
+import { createConfigurations } from "@webiny/api-headless-cms-utils-os/configurations.js";
 import { createTransformer } from "./transformations/index.js";
 import {
     createEntryLatestKeys,
