+++ packages/api-headless-cms-ddb-es/src/operations/entry/DdbEsGetUniqueFieldValues.ts
@@ -11,7 +11,7 @@ import type {
     SearchBody as OpenSearchSearchBody
 } from "@webiny/api-opensearch/types.js";
 import { CmsModelOpenSearchIndexProvider } from "~/features/CmsModelOpenSearchIndex/index.js";
-import { createConfigurations } from "~/configurations.js";
+import { createConfigurations } from "@webiny/api-headless-cms-utils-os/configurations.js";
 import { shouldIgnoreEsResponseError } from "@webiny/api-headless-cms-utils-os/operations/entry/elasticsearch/shouldIgnoreEsResponseError.js";
 import { CmsEntryOpenSearchBodyBuilder } from "@webiny/api-headless-cms-utils-os/exports/api/cms/opensearch.js";
