+++ packages/api-headless-cms-ddb-es/src/operations/entry/DdbEsListEntries.ts
@@ -13,7 +13,7 @@ import { createLimit, decodeCursor, encodeCursor } from "@webiny/api-opensearch"
 import type { OpenSearchSearchResponse } from "@webiny/api-opensearch/types.js";
 import { getTotalCount } from "@webiny/api-opensearch/types.js";
 import { CmsModelOpenSearchIndexProvider } from "~/features/CmsModelOpenSearchIndex/index.js";
-import { createConfigurations } from "~/configurations.js";
+import { createConfigurations } from "@webiny/api-headless-cms-utils-os/configurations.js";
 import { extractEntriesFromIndex } from "@webiny/api-headless-cms-utils-os/helpers/index.js";
 import { shouldIgnoreEsResponseError } from "@webiny/api-headless-cms-utils-os/operations/entry/elasticsearch/shouldIgnoreEsResponseError.js";
 import {
