+++ packages/api-headless-cms-pg-os/src/features/CmsModelOpenSearchIndex/feature.ts
@@ -1,11 +1,1 @@
-import { createFeature } from "@webiny/feature/api/index.js";
-import { DefaultCmsModelOpenSearchIndex } from "./DefaultCmsModelOpenSearchIndex.js";
-import { DefaultCmsModelOpenSearchIndexProvider } from "./CmsModelOpenSearchIndexProvider.js";
-
-export const CmsModelOpenSearchIndexFeature = createFeature({
-    name: "Cms/Model/OpenSearch/IndexFeature",
-    register: container => {
-        container.register(DefaultCmsModelOpenSearchIndex);
-        container.register(DefaultCmsModelOpenSearchIndexProvider).inSingletonScope();
-    }
-});
+export { CmsModelOpenSearchIndexFeature } from "@webiny/api-headless-cms-utils-os/features/CmsModelOpenSearchIndex/index.js";
