+++ packages/api-headless-cms-utils-os/src/features/CmsModelOpenSearchIndex/feature.ts
@@ -1,0 +1,11 @@
+import { createFeature } from "@webiny/feature/api/index.js";
+import { DefaultCmsModelOpenSearchIndex } from "./DefaultCmsModelOpenSearchIndex.js";
+import { DefaultCmsModelOpenSearchIndexProvider } from "./DefaultCmsModelOpenSearchIndexProvider.js";
+
+export const CmsModelOpenSearchIndexFeature = createFeature({
+    name: "Cms/Model/OpenSearch/IndexFeature",
+    register: container => {
+        container.register(DefaultCmsModelOpenSearchIndex);
+        container.register(DefaultCmsModelOpenSearchIndexProvider).inSingletonScope();
+    }
+});
