+++ packages/api-headless-cms-utils-os/src/features/CmsModelOpenSearchIndex/index.ts
@@ -9,3 +9,6 @@ export type {
     ICmsModelOpenSearchIndexProvider,
     ICmsModelOpenSearchIndexProviderResult
 } from "./CmsModelOpenSearchIndexProvider.js";
+export { DefaultCmsModelOpenSearchIndex } from "./DefaultCmsModelOpenSearchIndex.js";
+export { DefaultCmsModelOpenSearchIndexProvider } from "./DefaultCmsModelOpenSearchIndexProvider.js";
+export { CmsModelOpenSearchIndexFeature } from "./feature.js";
